Refactor key command handling by moving logic to a new composable. Removed KeyCommands.vue and implemented useShortcutKeys.ts for better modularity and reusability.
