+++ composables/useShortcutKeys.ts
@@ -1,0 +1,28 @@
+import { onKeyDown } from '@vueuse/core'
+import { useAuth } from '#imports'
+
+
+export function useShortcutKeys() {
+  const { status } = useAuth()
+  const dialog = useDialog()
+  
+  if (status.value === 'authenticated') {
+    onKeyDown(['t', 'l'], (event) => {
+      const activeElement = document.activeElement
+
+      if (activeElement && activeElement.className === 'v-overlay__content') return
+      if (activeElement && activeElement.tagName === 'INPUT') return
+      if (activeElement && activeElement.tagName === 'TEXTAREA') return
+
+      if (event.key === 't') {
+        dialog.value.open = true
+        dialog.value.page = 'todo'
+      }
+
+      if (event.key === 'l') {
+        dialog.value.open = true
+        dialog.value.page = 'list'
+      }
+    })
+  }
+}
